+++ vendor_ingestion_api/ingestion_api_model.py
@@ -618,7 +618,7 @@ class Emlockstatus_Ingest_DataParams(pydantic.BaseModel):
 
 class HyperLocalZoneSummaryCreate(pydantic.BaseModel):
     zone: str
-    rating: int
+    rating: float
 
 
 class ReviewsSentimentCreate(pydantic.BaseModel):
@@ -628,7 +628,7 @@ class ReviewsSentimentCreate(pydantic.BaseModel):
     store_code: str
     store_name: str
     reviewer_name: str
-    rating: int
+    rating: float
     review_comment: str
     tone_type: str
     review_date: datetime.datetime
